==> programs/2_Process_Summarize_ERA5.Rmd <==
---
title: "Process and Summarize ERA5 data"
author: "B Steele"
date: "2023-03-30"
output: html_document
---

```{r}
library(tidyverse)
library(ggthemes)
library(RcppRoll)
library(googledrive)
library(lubridate)

#point to directories
drv_dir = 'data/fromDrive/'
era_dir = 'data/era5_met/'

#authenticate drive
drive_auth()
```

# Purpose

This script creates 3, 5, 7 day summaries of the Copernicus ERA5 daily data for Yojoa for use in the downstream Secchi modeling efforts. The ERA5 data were acuuired in Google Earth Engine using the script `era5_download.js` found in the folder yojoa-rs-secchi/GEE_scripts.

### Suggested Citations:

-   Muñoz Sabater, J., (2019): ERA5-Land daily averaged data from 1981 to present. Copernicus Climate Change Service (C3S) Climate Data Store (CDS). (2023-03-30), [doi:10.24381/cds.68d2bb30](https://doi.org/10.24381/cds.68d2bb30)

# Download and Save Locally

GEE output are saved in Google Drive, first, let's move the ERA5 data to a local folder to save for posterity.

```{r}
#get did -- note, if you have a lot of files in your drive, you may want to define the folder before you search for the file.
dr_file = drive_ls(pattern = 'yojoa_era5_data.csv')
drive_download(as_id(dr_file$id), path = file.path(era_dir, dr_file$name), overwrite = T)
```

# Make summary files

### Load in ERA5 data and format date

```{r}
era5 = read.csv(file.path(era_dir, dr_file$name)) %>% 
  mutate(date = as.Date(substr(system.index, 1, 8), format = '%Y%m%d'))
```

### Calculate wind speed from u and v components

U and V are zonal and meridional winds characteristics. Wind speed is calculated using the Pythagorean theorem, where wind speed is equal to the square root of the sum of the squares of the u and v characteristics. For more information about this, see [NASA's description and U and V components](https://disc.gsfc.nasa.gov/information/data-in-action?title=Derive%20Wind%20Speed%20and%20Direction%20With%20MERRA-2%20Wind%20Components).

```{r}
era5$wind_speed_mps = sqrt(era5$u_component_of_wind_10m^2 + era5$v_component_of_wind_10m^2)
```

### Select the parameters we're interested in and rename them with units

```{r}
era5_sub = era5 %>% 
  select(date, 
         solar_rad_Jpm2 = surface_solar_radiation_downwards_sum,
         precip_m = total_precipitation_sum,
         air_temp_degK = temperature_2m,
         wind_speed_mps)
```

### Summarize data in 1,3,5,7 day summaries

Previous day's weather
```{r}
previous = era5_sub %>% 
  mutate(date = date + days(1)) %>% 
  rename_with(~paste0(., '_prev'),
              c('solar_rad_Jpm2', 'precip_m', 'air_temp_degK', 'wind_speed_mps')) %>% 
  mutate(solar_rad_KJpm2_prev = solar_rad_Jpm2_prev*0.001) %>% 
  select(-solar_rad_Jpm2_prev)
```


Previous 3 days summary

```{r}
threeday = as.data.frame(era5_sub$date[3:nrow(era5_sub)])
colnames(threeday) = 'date'
threeday$tot_sol_rad_KJpm2_3 = roll_sum(x = era5_sub$solar_rad_Jpm2, align = 'right', 3)*0.001
threeday$mean_temp_degK_3 = roll_mean(x  = era5_sub$air_temp_degK, align = 'right', 3)
threeday$min_temp_degK_3 = roll_min(x  = era5_sub$air_temp_degK, align = 'right', 3)
threeday$max_temp_degK_3 = roll_max(x  = era5_sub$air_temp_degK, align = 'right', 3)
threeday$tot_precip_m_3 = roll_sum(x  = era5_sub$precip_m, align = 'right', 3)
threeday$mean_wind_mps_3 = roll_mean(x  = era5_sub$wind_speed_mps, align = 'right', 3)
```

Previous 5 days summary

```{r}
fiveday = as.data.frame(era5_sub$date[5:nrow(era5_sub)])
colnames(fiveday) = 'date'
fiveday$tot_sol_rad_KJpm2_5 = roll_sum(x = era5_sub$solar_rad_Jpm2, align = 'right', 5)*0.001
fiveday$mean_temp_degK_5 = roll_mean(x  = era5_sub$air_temp_degK, align = 'right', 5)
fiveday$min_temp_degK_5 = roll_min(x  = era5_sub$air_temp_degK, align = 'right', 5)
fiveday$max_temp_degK_5 = roll_max(x  = era5_sub$air_temp_degK, align = 'right', 5)
fiveday$tot_precip_m_5 = roll_sum(x  = era5_sub$precip_m, align = 'right', 5)
fiveday$mean_wind_mps_5 = roll_mean(x  = era5_sub$wind_speed_mps, align = 'right', 5)
```

Seven-day summary

```{r}
sevenday = as.data.frame(era5_sub$date[7:nrow(era5_sub)])
colnames(sevenday) = 'date'
sevenday$tot_sol_rad_KJpm2_7 = roll_sum(x = era5_sub$solar_rad_Jpm2, align = 'right', 7)*0.001
sevenday$mean_temp_degK_7 = roll_mean(x  = era5_sub$air_temp_degK, align = 'right', 7)
sevenday$min_temp_degK_7 = roll_min(x  = era5_sub$air_temp_degK, align = 'right', 7)
sevenday$max_temp_degK_7 = roll_max(x  = era5_sub$air_temp_degK, align = 'right', 7)
sevenday$tot_precip_m_7 = roll_sum(x  = era5_sub$precip_m, align = 'right', 7)
sevenday$mean_wind_mps_7 = roll_mean(x  = era5_sub$wind_speed_mps, align = 'right', 7)
```

Join them together, adjust date so that all cumulative data are prior to the Landsat/Secchi observation, and save the file

```{r}
era5_summary = full_join(threeday, fiveday) %>% 
  full_join(., sevenday) %>% 
  mutate(date = date + days(1)) %>% 
  full_join(., previous) %>% 
  arrange(date)
```

Save the file for use in modeling efforts

```{r}
write.csv(era5_summary, file.path(era_dir, 'Yojoa_1-3-5-7day_summary_ERA5.csv'), row.names = F)
```
