==> programs/4_regional_handoff_calcs.Rmd <==
---
title: "Landsat 4-7 reflectance handoff to Landsat 8-9"
author: "B Steele"
date: "2023-03-28"
output: html_document
---

```{r}
library(tidyverse)
```

# Purpose

Landsat 4-7 and 8-9 surface reflectance data go through two different atmospheric corrections ([LEDAPS](https://pubs.er.usgs.gov/publication/ofr20131057) and [LaSRC](https://ntrs.nasa.gov/api/citations/20190001670/downloads/20190001670.pdf)). Additionally, each band wavelength can vary between missions. This script uses an adapted version of the methods in Topp, et al. 2021 to correct for each satellite handoff, correction to LS 7 values.

# Read in filtered data

These data were filtered such that dswe1 \>= 10, image quality was \>= 7, and all median Rrs values between -0.01 and 0.2 in the program `3_regionalRS_collate_harmonize.Rmd`

```{r}
filtered = read.csv('data/upstreamRS/HON_GUAT_ELSAL_regional_LandsatC2_SRST_filtered_v2023-03-31.csv') %>% 
  mutate(date = as.Date(date),
         rowid = as.integer(rowid))
```

# Build functions for each overlap

Here, we use the Topp et al version of the functions to calculate handoff coefficients. In this adaptation, I adjusted the filter to align with deployment and decommission dates for each sensor. Additionally, I added code to display each polynomial regression and provide some summary information for each handoff per band. Because the number of scenes was low for LS 9 (n \< 50), I adjusted the quantiles to 0.05 increments: `quantile(., seq(.05,.95, .05))` as to not oversample the data.

```{r}
## Create polynomial functions based on the 1-99th percentile of each sensor
## for overlapping periods
lm75 <- function(band){
  print(paste0(band, ' model summary'))
  y <- filtered %>% 
    filter(date > as.Date('1999-04-15'), date < as.Date('2013-06-05'), mission == 'LANDSAT_7') 
  print(paste0('LS7 scenes: ', length(unique(y$system.index)), ' values: ', nrow(y)))
  y <- y %>% 
    .[,band] %>% 
    quantile(., seq(.01,.99, .01))
  
  x = filtered %>% 
    filter(date > as.Date('1999-04-15'), date < as.Date('2013-06-05'), mission == 'LANDSAT_5') 
  print(paste0('LS5 scenes: ', length(unique(x$system.index)), ' values: ', nrow(x)))
  x = x %>% 
    .[,band] %>% 
    quantile(., seq(.01,.99, .01))
  
  lm <- lm(y~poly(x, 2, raw = T))
  print(summary(lm))
  plot(y~x, 
       main = paste0(band, ' LS5-7 handoff'), 
       ylab = '0.01 Quantile Values for LS7 Rrs', 
       xlab = '0.01 Quantile Values for LS5 Rrs')
  lines(sort(x),
        fitted(lm)[order(x)],
        col = "red",
        type = "l")
  
  df <- tibble(band = band, intercept = lm$coefficients[[1]], B1 = lm$coefficients[[2]], B2 = lm$coefficients[[3]])
  return(df)
}
lm78 <- function(band){
  print(paste0(band, ' model summary'))
  y <- filtered %>% 
    filter(date > as.Date('2013-02-11'), date < as.Date('2022-04-16'), mission == 'LANDSAT_7') 
  print(paste0('LS7 scenes: ', length(unique(y$system.index)), ' values: ', nrow(y)))
  y <- y %>% 
    .[,band] %>% 
    quantile(., seq(.01,.99, .01))

  x = filtered %>% 
    filter(date > as.Date('2013-02-11'), date < as.Date('2022-04-16'), mission == 'LANDSAT_8') 
  print(paste0('LS8 scenes: ', length(unique(x$system.index)), ' values: ', nrow(x)))
  x = x %>% 
    .[,band] %>% 
    quantile(., seq(.01,.99, .01))

  lm <- lm(y~poly(x, 2, raw = T))
  print(summary(lm))
  plot(y~x, main = paste0(band, ' LS7-8 handoff'), 
       ylab = '0.01 Quantile Values for LS7 Rrs', 
       xlab = '0.01 Quantile Values for LS8 Rrs')
  lines(sort(x),
        fitted(lm)[order(x)],
        col = "red",
        type = "l")
  
  df <- tibble(band = band, intercept = lm$coefficients[[1]], B1 = lm$coefficients[[2]], B2 = lm$coefficients[[3]])
  return(df)
}
lm79 <- function(band){
  print(paste0(band, ' model summary'))
  y <- filtered %>% 
    filter(date > as.Date('2021-09-27'), date < as.Date('2022-04-16'), mission == 'LANDSAT_7') 
  print(paste0('LS7 scenes: ', length(unique(y$system.index)), ' values: ', nrow(y)))
  y <- y %>% 
    .[,band] %>% 
    quantile(., seq(.05,.95, .05))

  x = filtered %>% 
    filter(date > as.Date('2021-09-27'), date < as.Date('2022-04-16'), mission == 'LANDSAT_9') 
  print(paste0('LS9 scenes: ', length(unique(x$system.index)), ' values: ', nrow(x)))
  x = x %>% 
    .[,band] %>% 
    quantile(., seq(.05,.95, .05))

  lm <- lm(y~poly(x, 2, raw = T))
  print(summary(lm))
  plot(y~x, main = paste0(band, ' LS7-9 handoff'), 
       ylab = '0.05 Quantile Values for LS7 Rrs', 
       xlab = '0.05 Quantile Values for LS9 Rrs')
  lines(sort(x),
        fitted(lm)[order(x)],
        col = "red",
        type = "l")
  
  df <- tibble(band = band, intercept = lm$coefficients[[1]], B1 = lm$coefficients[[2]], B2 = lm$coefficients[[3]])
  return(df)
}

```

## Run the functions

Now, we can run the functions and look at the data by correction and band

```{r}
## Run the functions and look at the resulting corrections
bands <-  c('med_Blue', 'med_Green', 'med_Red', 'med_Nir', 'med_Swir1','med_Swir2')
```

### Landsat 5 to 7

```{r}
funcs.5 <- bands %>% map_dfr(lm75)%>% mutate(SatCorr = 'LANDSAT_5')
```

### Landsat 8 to 7

```{r}
funcs.8 <- bands %>% map_dfr(lm78) %>% mutate(SatCorr = 'LANDSAT_8')

```

### Landsat 9 to 7

```{r}
funcs.9 <- bands %>% map_dfr(lm79)%>% mutate(SatCorr = 'LANDSAT_9')
```

### Join all coefficients

```{r}
handoffs = full_join(funcs.5, funcs.8) %>% 
  full_join(., funcs.9)
```

## Landsat 4 -7 Handoff

While there is technically overlap between the LS4 and LS7 mission (1999-2001), there are no scenes from LS4 that are available for that time for the region. The final image in the LS4 stack is dated:

```{r}
filtered %>% filter(mission == 'LANDSAT_4') %>% distinct(date) %>% arrange(date) %>% tail(1)
```

## Save handoff coefficients

```{r}
write.csv(handoffs, file.path('data/upstreamRS/', paste0('yojoa_regional_handoff_coefficients_v', Sys.Date(),'.csv')), row.names = F)
```
