==> programs/11_stepwise_regression_method.Rmd <==
---
title: "Stepwise Regression Method"
author: "B Steele"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(tidyverse)
library(caret)
library(leaps)
library(ggpmisc)
library(ggthemes)

match_dir = 'data/matchups/'
rs_dir = 'data/upstreamRS/'
is_dir = 'data/in-situ/'

# Set seed for reproducibility
set.seed(799)
```

# Purpose

This script uses stepwise regression as an alternative modeling approach for the Lake Yojoa Secchi data.

## Load data

```{r}
match5 = read.csv(file.path(match_dir, 'fiveDay_LS-Secchi_matchups_n237.csv'))
match71 = read.csv(file.path(match_dir, 'multiDay_LS-Secchi_matchups_n238.csv'))

stack = read.csv(file.path(rs_dir, 'yojoa_corr_rrs_met_scaled_v2023-06-15.csv'))
secchi = read.csv(file.path(is_dir, 'Secchi_completedataset.csv')) %>% 
  mutate(secchi = as.numeric(secchi),
         date = mdy(date)) %>% 
  filter(!is.na(secchi)) %>% 
  mutate(location = gsub(' ', '', location))
```

## Prep data

```{r}
prepData = function(df) {
  #make a rowid column
  df_prep = df %>% 
    rowid_to_column() %>% 
    mutate(secchi = as.numeric(secchi)) %>% #there's one wonky value in here with two decimal points... dropping from this analysis
    filter(!is.na(secchi))
}

match5 = prepData(match5)
match71 = prepData(match71)
stack = stack %>% 
  filter(date < as.Date('2023-01-01'))
```

Make train/test/validation sets

70% train, 30% test. We use a cross-validation technique in the stepwise regression model that acts as a train/test (on the 70% data), and use the remaining 30% as a holdout for validation.

```{r}
##Pull 30% as holdout test data
test_match5 <- match5 %>%
  sample_frac(0.3) 

## Remove holdout data
train_match5 <- match5 %>% filter(!rowid %in% test_match5$rowid) 

test_match71 <- match71 %>% 
  sample_frac(0.3)

train_match71 <- match71 %>% filter(!rowid %in% test_match71$rowid) 

```

Point to the features for model and subset data for those vars. We'll use scaled variables in this method since linear-based models are susceptible to differing variable ranges.

```{r}
vars_3m = c('secchi',
         'med_Blue_corr_scaled', 'med_Red_corr_scaled', 'med_Green_corr_scaled', 
         'med_Nir_corr_scaled',
         'RN_scaled', 'BG_scaled', 'RB_scaled', 'GB_scaled', 
         "tot_sol_rad_KJpm2_3_scaled", "max_temp_degK_3_scaled", "mean_temp_degK_3_scaled", "min_temp_degK_3_scaled",
         "tot_precip_m_3_scaled", "mean_wind_mps_3_scaled")

vars_5m = c('secchi',
         'med_Blue_corr_scaled', 'med_Red_corr_scaled', 'med_Green_corr_scaled', 
         'med_Nir_corr_scaled',
         'RN_scaled', 'BG_scaled', 'RB_scaled', 'GB_scaled', 
         "tot_sol_rad_KJpm2_5_scaled", "max_temp_degK_5_scaled", "mean_temp_degK_5_scaled", "min_temp_degK_5_scaled",
         "tot_precip_m_5_scaled", "mean_wind_mps_5_scaled")

vars_51m = c('secchi',
         'med_Blue_corr_scaled', 'med_Red_corr_scaled', 'med_Green_corr_scaled', 
         'med_Nir_corr_scaled',
         'RN_scaled', 'BG_scaled', 'RB_scaled', 'GB_scaled', 
         "tot_sol_rad_KJpm2_5_scaled", "max_temp_degK_5_scaled", "mean_temp_degK_5_scaled", "min_temp_degK_5_scaled",
         "tot_precip_m_5_scaled", "mean_wind_mps_5_scaled", 
         "precip_m_prev_scaled", "air_temp_degK_prev_scaled", "wind_speed_mps_prev_scaled")

vars_71m = c('secchi',
         'med_Blue_corr_scaled', 'med_Red_corr_scaled', 'med_Green_corr_scaled', 
         'med_Nir_corr_scaled',
         'RN_scaled', 'BG_scaled', 'RB_scaled', 'GB_scaled', 
         "tot_sol_rad_KJpm2_7_scaled", "max_temp_degK_7_scaled", "mean_temp_degK_7_scaled", "min_temp_degK_7_scaled",
         "tot_precip_m_7_scaled", "mean_wind_mps_7_scaled", 
         "precip_m_prev_scaled", "air_temp_degK_prev_scaled", "wind_speed_mps_prev_scaled")

#subset for only vars for modeling
train_match_5d_3m = train_match5 %>% select(all_of(vars_3m))
train_match_5d_5m = train_match5 %>% select(all_of(vars_5m))
train_match_5d_51m = train_match5 %>% select(all_of(vars_51m))
train_match_5d_71m = train_match5 %>% select(all_of(vars_71m))

train_match_71d_3m = train_match71 %>% select(all_of(vars_3m))
train_match_71d_5m = train_match71 %>% select(all_of(vars_5m))
train_match_71d_51m = train_match71 %>% select(all_of(vars_51m))
train_match_71d_71m = train_match71 %>% select(all_of(vars_71m))
```

And run the model

```{r, eval = F}
# Set up repeated k-fold cross-validation
train.control <- trainControl(method = "cv", number = 5)

# Train the models
step.model.5d3m <- train(secchi ~ ., data = train_match_5d_3m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.5d3m, file = 'data/models/stepreg_5d_3m.RData')

step.model.5d5m <- train(secchi ~ ., data = train_match_5d_5m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.5d5m, file = 'data/models/stepreg_5d_5m.RData')

step.model.5d51m <- train(secchi ~ ., data = train_match_5d_51m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.5d51m, file = 'data/models/stepreg_5d_51m.RData')

step.model.5d71m <- train(secchi ~ ., data = train_match_5d_71m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.5d71m, file = 'data/models/stepreg_5d_71m.RData')

step.model.71d3m <- train(secchi ~ ., data = train_match_71d_3m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.71d3m, file = 'data/models/stepreg_71d_3m.RData')

step.model.71d5m <- train(secchi ~ ., data = train_match_71d_5m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )
save(step.model.71d5m, file = 'data/models/stepreg_71d_5m.RData')

step.model.71d51m <- train(secchi ~ ., data = train_match_71d_51m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )

save(step.model.71d51m, file = 'data/models/stepreg_71d_51m.RData')

step.model.71d71m <- train(secchi ~ ., data = train_match_71d_71m,
                    method = "leapBackward", 
                    tuneGrid = data.frame(nvmax = 1:13),
                    trControl = train.control
                    )

save(step.model.71d71m, file = 'data/models/stepreg_71d_71m.RData')

```

Load models and see results

```{r}
load('data/models/stepreg_5d_3m.RData')
load('data/models/stepreg_5d_5m.RData')
load('data/models/stepreg_5d_51m.RData')
load('data/models/stepreg_5d_71m.RData')

load('data/models/stepreg_71d_3m.RData')
load('data/models/stepreg_71d_5m.RData')
load('data/models/stepreg_71d_51m.RData')
load('data/models/stepreg_71d_71m.RData')

step.model.5d3m$results
step.model.5d5m$results
step.model.5d51m$results
step.model.5d71m$results

step.model.71d3m$results
step.model.71d5m$results
step.model.71d51m$results
step.model.71d71m$results

```

Looks like the only model to really dig in with is the 5-day window and the 7/1 day window with the 5 day met. Let's look at the hold out data to see what they look like before we go any further. First, the 7/1 day window:

```{r}
test_match71$predicted = predict.train(step.model.71d5m, test_match71)

ggplot(test_match71, aes(x = secchi, y = predicted)) + 
  geom_point() +
  geom_abline(color = 'grey', lty = 2) + 
  # coord_cartesian(xlim = c(0, 6.5),
  #                 ylim = c(0,6.5)) +
  stat_poly_eq(aes(label = paste(after_stat(adj.rr.label))),
               formula = y~x, 
               parse = TRUE, 
               label.y = Inf, 
               vjust = 1.3) +
  labs(title = 'Yojoa Secchi Stepwise Regression Test Data\n7/1 day matchups, band and 5-day met summaries', 
       subtitle = 'moderate data stringency\ngrey dashed line is 1:1', 
       x = 'Actual Secchi (m)', 
       y = 'Predicted Secchi (m)')  +
  theme_bw()+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))

test_match71 %>% 
  filter(predicted>0) %>% 
  ggplot(., aes(x = secchi, y = predicted)) + 
  geom_point() +
  geom_abline(color = 'grey', lty = 2) + 
  coord_cartesian(xlim = c(0, 6.5),
                  ylim = c(0,6.5)) +
  stat_poly_eq(aes(label = paste(after_stat(adj.rr.label))),
               formula = y~x, 
               parse = TRUE, 
               label.y = Inf, 
               vjust = 1.3) +
  labs(title = 'Yojoa Secchi Stepwise Regression Test Data\n7/1 day matchups, band and 5-day met summaries', 
       subtitle = 'moderate data stringency\ngrey dashed line is 1:1', 
       x = 'Actual Secchi (m)', 
       y = 'Predicted Secchi (m)')  +
  theme_bw()+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))
```

And look at the 5-day window

```{r}
test_match5$predicted = predict.train(step.model.5d5m, test_match5)

ggplot(test_match5, aes(x = secchi, y = predicted)) + 
  geom_point() +
  geom_abline(color = 'grey', lty = 2) + 
  # coord_cartesian(xlim = c(0, 6.5),
  #                 ylim = c(0,6.5)) +
  stat_poly_eq(aes(label = paste(after_stat(adj.rr.label))),
               formula = y~x, 
               parse = TRUE, 
               label.y = Inf, 
               vjust = 1.3) +
  labs(title = 'Yojoa Secchi Stepwise Regression Test Data\nfive day matchups, band and 5-day met summaries', 
       subtitle = 'moderate data stringency\ngrey dashed line is 1:1', 
       x = 'Actual Secchi (m)', 
       y = 'Predicted Secchi (m)')  +
  theme_bw()+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))

test_match5 %>% 
  filter(predicted > 0) %>% 
  ggplot(., aes(x = secchi, y = predicted)) + 
  geom_point() +
  geom_abline(color = 'grey', lty = 2) + 
  coord_cartesian(xlim = c(0, 6.5),
                  ylim = c(0,6.5)) +
  stat_poly_eq(aes(label = paste(after_stat(adj.rr.label))),
               formula = y~x, 
               parse = TRUE, 
               label.y = Inf, 
               vjust = 1.3) +
  labs(title = 'Yojoa Secchi Stepwise Regression Test Data\nfive day matchups, band and 5-day met summaries', 
       subtitle = 'moderate data stringency\ngrey dashed line is 1:1', 
       x = 'Actual Secchi (m)', 
       y = 'Predicted Secchi (m)')  +
  theme_bw()+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))
```

Okay, well those negative values suck, so let's take a peek at the timeseries for the 7/1 window, since the 5 day has egregious Secchi estimates for some out-of-range features.

```{r}
stack = stack %>% 
  mutate(secchi = predict.train(object = step.model.71d5m, newdata = stack)) %>% 
  mutate(date = ymd(date)) 
secchi = secchi %>% 
  mutate(mission = 'Measured') 

situ_stack <- full_join(stack, secchi) %>% 
  select(date, location, secchi, mission) 
```

```{r}
plotRecordBySite = function(site) {
  ggplot(situ_stack %>%
           filter(location == site), aes(x = date, y = secchi, color = mission,
                                        shape = mission)) + 
    geom_point() + 
    labs(title = paste0('Yojoa Secchi Historical Record - Site ', site),
       subtitle = '7/1 day window, 5 day met\nmoderate data stringency\nstepwise regression',
         y = 'Secchi (m)',
         color = 'data source', shape = 'data source') +
    scale_color_manual(values = c('grey10','grey30','grey50','grey70','blue')) + 
    theme_few() +
    theme(legend.position = c(0.8,0.8)) + 
    scale_shape_manual(values = c(19,19,19,19,1)) +
    scale_y_continuous(limits = c(0, max(situ_stack$secchi)), breaks = seq(0, max(situ_stack$secchi), 2)) +
    scale_x_date(limits = c(min(situ_stack$date), max(situ_stack$date))) +
    theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
          plot.subtitle = element_text(hjust = 0.5))
}

map(sort(unique(situ_stack$location)), plotRecordBySite)
```

### Look at recent data per location

```{r}
plotRecentBySite = function(site) {
  ggplot(situ_stack %>%
           filter(location == site), aes(x = date, y = secchi, color = mission,
                                        shape = mission)) + 
    geom_point() + 
    labs(title = paste0('Yojoa Secchi 2018-2022 - Site ', site),
       subtitle = '7/1 day window, 5 day met\nmoderate data stringency\nstepwise regression',
         y = 'Secchi (m)',
         color = 'data source', shape = 'data source') +
    scale_color_manual(values = c('grey10','grey30','grey50','grey70','blue')) + 
    theme_few() +
    theme(legend.position = c(0.8,0.8)) + 
    scale_shape_manual(values = c(19,19,19,19,1)) +
    scale_y_continuous(limits = c(0, max(situ_stack$secchi)), breaks = seq(0, max(situ_stack$secchi), 2)) +
    scale_x_date(limits = c(ymd('2018-01-01'), max(situ_stack$date))) +
    theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
          plot.subtitle = element_text(hjust = 0.5))
}

map(sort(unique(situ_stack$location)), plotRecentBySite)
```


### Whole lake secchi dynamics

While there is plenty of variability across the lake, let's summarize to a single value per date, since not all sites have the same density of record. Since there are a few oddballs in here (both in terms of measured and estimated), we'll use the median Secchi across all sites.

```{r}
lake_avg <- situ_stack %>%
  group_by(date,mission) %>%
  summarize(across(where(is.numeric),median))

ggplot(lake_avg, aes(x = date, y = secchi, color = mission, shape = mission)) + 
  geom_point() + 
  scale_color_manual(values = c('grey10','grey30','grey50','grey70','blue')) + 
  labs(title = 'Yojoa Secchi Historical Record\nwhole-lake average',
       subtitle = '7/1 day window, 5 day met\nmoderate data stringency\nstepwise regression',
       y = 'median Secchi (m)',
       color = 'data source', shape = 'data source') +
  theme_few() +
  theme(legend.position = c(0.8,0.8)) + 
  scale_shape_manual(values = c(19,19,19,19,1)) +
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))


```

## And look at overlapping times to see how we're doing

### 2006

```{r}
ggplot(lake_avg, aes(x = date, y = secchi, color = mission,shape = mission)) + 
  geom_point() + 
  scale_color_manual(values = c('grey10','grey30','grey50','grey70','blue')) + 
  theme_few() +
  labs(title = 'Yojoa Secchi 2006\nwhole-lake average',
       subtitle = '7/1 day window, 5 day met\nmoderate data stringency\nstepwise regression',
       y = 'median Secchi (m)',
       color = 'data source', shape = 'data source') +
  theme(legend.position = c(0.8,0.8)) + 
  scale_shape_manual(values = c(19,19,19,19,1)) +
  scale_x_date(limits = c(as.Date('2006-01-01'), as.Date('2006-12-31')))+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))

```

### Recent

```{r}
ggplot(lake_avg, aes(x = date, y = secchi, color = mission, shape = mission)) + 
  geom_point() + 
  scale_color_manual(values = c('grey10','grey30','grey50','grey70','blue')) + 
  theme_few() +
  labs(title = 'Yojoa Secchi 2018-2022\nwhole-lake average',
       subtitle = '7/1 day window, 5 day met\nmoderate data stringency\nstepwise regression',
       y = 'median Secchi (m)',
       color = 'data source', shape = 'data source') +
  theme(legend.position = c(0.8,0.8)) + 
  scale_shape_manual(values = c(19,19,19,19,1)) +
  scale_x_date(limits = c(as.Date('2018-01-01'), as.Date('2023-01-01')))+
  theme(plot.title = element_text(hjust = 0.5, face = 'bold'),
        plot.subtitle = element_text(hjust = 0.5))

```

### Look at features in selected stepwise regression model

```{r}
varImp(step.model.71d5m)
```
